+++ src/components/block-tunes/block-tune-delete.ts
@@ -4,7 +4,7 @@
  * @copyright <CodeX Team> 2018
  */
 import { API, BlockTune, PopoverItem } from '../../../types';
-import $ from '../dom';
+import { IconCross } from '@codexteam/icons';
 
 /**
  *
@@ -36,8 +36,7 @@ export default class DeleteTune implements BlockTune {
    */
   public render(): PopoverItem {
     return {
-      // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-      icon: $.svg('cross', 14, 14).outerHTML,
+      icon: IconCross,
       label: this.api.i18n.t('Delete'),
       name: 'delete',
       confirmation: {