+++ src/components/block-tunes/block-tune-move-down.ts
@@ -4,9 +4,10 @@
  * @copyright <CodeX Team> 2018
  */
 
-import $ from '../dom';
 import { API, BlockTune, PopoverItem } from '../../../types';
 import Popover from '../utils/popover';
+import { IconChevronDown } from '@codexteam/icons';
+
 
 /**
  *
@@ -45,8 +46,7 @@ export default class MoveDownTune implements BlockTune {
    */
   public render(): PopoverItem {
     return {
-      // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-      icon: $.svg('arrow-down', 14, 14).outerHTML,
+      icon: IconChevronDown,
       label: this.api.i18n.t('Move down'),
       onActivate: (item, event): void => this.handleClick(event),
       name: 'move-down',