+++ src/components/inline-tools/inline-tool-bold.ts
@@ -1,5 +1,5 @@
-import $ from '../dom';
 import { InlineTool, SanitizerConfig } from '../../../types';
+import { IconBold } from '@codexteam/icons';
 
 /**
  * Bold Tool
@@ -61,8 +61,7 @@ export default class BoldInlineTool implements InlineTool {
     this.nodes.button = document.createElement('button') as HTMLButtonElement;
     this.nodes.button.type = 'button';
     this.nodes.button.classList.add(this.CSS.button, this.CSS.buttonModifier);
-    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-    this.nodes.button.appendChild($.svg('bold', 12, 14));
+    this.nodes.button.innerHTML = IconBold;
 
     return this.nodes.button;
   }