+++ src/components/inline-tools/inline-tool-italic.ts
@@ -1,5 +1,5 @@
-import $ from '../dom';
 import { InlineTool, SanitizerConfig } from '../../../types';
+import { IconItalic } from '@codexteam/icons';
 
 /**
  * Italic Tool
@@ -61,8 +61,7 @@ export default class ItalicInlineTool implements InlineTool {
     this.nodes.button = document.createElement('button') as HTMLButtonElement;
     this.nodes.button.type = 'button';
     this.nodes.button.classList.add(this.CSS.button, this.CSS.buttonModifier);
-    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-    this.nodes.button.appendChild($.svg('italic', 4, 11));
+    this.nodes.button.innerHTML = IconItalic;
 
     return this.nodes.button;
   }