+++ src/components/inline-tools/inline-tool-link.ts
@@ -1,9 +1,8 @@
 import SelectionUtils from '../selection';
-
-import $ from '../dom';
 import * as _ from '../utils';
-import { API, InlineTool, SanitizerConfig } from '../../../types';
+import { InlineTool, SanitizerConfig, API } from '../../../types';
 import { Notifier, Toolbar, I18n, InlineToolbar } from '../../../types/api';
+import { IconLink, IconUnlink } from '@codexteam/icons';
 
 /**
  * Link Tool
@@ -123,10 +122,8 @@ export default class LinkInlineTool implements InlineTool {
     this.nodes.button = document.createElement('button') as HTMLButtonElement;
     this.nodes.button.type = 'button';
     this.nodes.button.classList.add(this.CSS.button, this.CSS.buttonModifier);
-    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-    this.nodes.button.appendChild($.svg('link', 14, 10));
-    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-    this.nodes.button.appendChild($.svg('unlink', 15, 11));
+
+    this.nodes.button.innerHTML = IconLink;
 
     return this.nodes.button;
   }
@@ -194,6 +191,7 @@ export default class LinkInlineTool implements InlineTool {
     const anchorTag = this.selection.findParentTag('A');
 
     if (anchorTag) {
+      this.nodes.button.innerHTML = IconUnlink;
       this.nodes.button.classList.add(this.CSS.buttonUnlink);
       this.nodes.button.classList.add(this.CSS.buttonActive);
       this.openActions();
@@ -207,6 +205,7 @@ export default class LinkInlineTool implements InlineTool {
 
       this.selection.save();
     } else {
+      this.nodes.button.innerHTML = IconLink;
       this.nodes.button.classList.remove(this.CSS.buttonUnlink);
       this.nodes.button.classList.remove(this.CSS.buttonActive);
     }