+++ webpack.config.js
@@ -91,15 +91,7 @@ module.exports = (env, argv) => {
           use: [
             'postcss-loader',
           ],
-        },
-        {
-          test: /\.(svg)$/,
-          use: [
-            {
-              loader: 'raw-loader',
-            },
-          ],
-        },
+        }
       ],
     },